bazel.sh: Use location of bazel.sh to locate TOP.
The bazel.sh wrapper script uses the existence of a file to check if
a given directory is TOP. Currently, it uses envsetup.sh. To allow for
standalone builds where envsetup.sh is not present, this changes the
script to use itself as the file to check for, since it is guaranteed to
be present.

Test: build/bazel/bazel.sh info
Change-Id: I05944d8110aea130e2a1b105fb54df35cd8e9bff