bp2build: upload cc_object macro.
This macro help to manage default attribute semantics, and can also
handle attributes not supported in the underlying rule class (see code
for example with the native_bridge_supported attribute, which is thrown
away for now.).

It also globs up the header files into hdrs from the directories
where sources live in to converge with Soong's behavior of allowing
.c/.cpp files to depend on .h files in the current directory, while
still allowing Bazel sandboxing to work correctly.

Test: GENERATE_BAZEL_FILES; sync; build crt_beginso1
Change-Id: I0b51e0735f82f98ec6b188ab558995d1d048bd04