bazel.sh: Read bazelrc location from environment.
This allows builds using bazel.sh to specify the bazelrc to use in the
ANDROID_BAZELRC_PATH environment variable. If that environment variable
is a relative path, it will be resolved relative to the TOP directory.
If ANDROID_BAZELRC_PATH specifies a directory, one of darwin.bazelrc or
linux.bazelrc will be used from that directory, based on the OS on which
the build is running.

Test: build with bazel.sh in standalone build
Change-Id: Ifd74e60995b80290ac7245c21abd2cc742b8ed07