cc_object: create a rule that actually returns objects
Instead of doing a full link with cc_library, this CL replaces the
native.cc_library call to an actual rule that calls into cc_common and
runs a compile, not link, on the input files, and returns a DefaultInfo
containing only the objects in compilation outputs.

The previous implementation would have built a static archive, which was
wrong.

This requires https://github.com/bazelbuild/bazel/pull/13155 to be
merged first, since the cc_common.compile API today doesn't accept .S
files as inputs, which are valid assembly files.

$ bazel-dev --bazelrc=build/bazel/linux.bazelrc cquery 'kind(_cc_object, //bionic/libc:all)' --output=starlark --starlark:expr="str(target.label) + ' ' + ' '.join([f.basename for f in target.files.to_list()])"

//bionic/libc:crtbegin_dynamic crtbegin.o crtbrand.o
//bionic/libc:crtbegin_dynamic1 crtbegin.o
//bionic/libc:crtbegin_so crtbegin_so.o crtbrand.o
//bionic/libc:crtbegin_so1 crtbegin_so.o
//bionic/libc:crtbegin_static crtbegin.o crtbrand.o
//bionic/libc:crtbegin_static1 crtbegin.o
//bionic/libc:crtbrand crtbrand.o
//bionic/libc:crtend_android crtend.o
//bionic/libc:crtend_so crtend_so.o
//bionic/libc:libseccomp_gen_syscall_nrs_arm gen_syscall_nrs.o
//bionic/libc:libseccomp_gen_syscall_nrs_arm64 gen_syscall_nrs.o
//bionic/libc:libseccomp_gen_syscall_nrs_x86 gen_syscall_nrs_x86.o
//bionic/libc:libseccomp_gen_syscall_nrs_x86_64 gen_syscall_nrs_x86_64.o

Test: build/bazel/scripts/milestone-2/demo.sh full
Test: go tests
Change-Id: If1944afb66c756a656399972711491ba489410ec