Handle no-srcs no-deps for cc_library_static.bzl
This creates a "safeguard" rule and target which only operates in the
no-srcs no-deps case of the main target, by invoking llvm-ar to create a
stub file. Otherwise, this safeguard is a no-op.

This is required for mixed build support of such no-srcs no-deps library
modules. Such cases typically exist when a module is defined with only
srcs/deps in some subset of architectures, but empty in others.

Test: USE_BAZEL_ANALYSIS=1 m libc
Test: USE_BAzEL_ANALYSIS=1 m runtime-module-sdk
Change-Id: Ibc6e1f9ed5ae49f1d60d92f2944ba789b890405b