bp2build: check in cc_library rules and unit test.
This change adds the Starlark code used by the bp2build-generated
targets for cc_library related types.

It also adds a sh_test unit test that analyzes the runfiles outputs of a
full cc_library generated by bp2build, starting with ld-android. It
analyzes that the shared and static libraries are of the right
filetypes, and that the shared library exports an expected set of
symbols.

This CL checks in cc_shared_library from rules_cc HEAD, without any
local changes for Roboleaf.

Test: TH
Test: bp2build, bazel test //build/bazel/tests/...

Change-Id: I4deb3fc3160e393b03b4bd550777146d6cad6de8