Update ci scripts to avoid bp2build sync
With recent changes to mixed builds, bp2build generation is integrated
into the build, and thus separate bp2build invocation is not required
for mixed builds. Additionally, bp2build sync is not required for
running builds with tools/bazel, as long as --package_path is set to
out/soong/workspace.

We will likely want to follow up and handle bp2build-detection and
setting --package_path inline as part of tools/bazel.

Test: Ran all ci scripts manually
Change-Id: Iae7a357d46543eb3f667b31ce496e6e0f6a1fa50