Add support for dynamic_deps in cc_library_static.
libc depends on libdl via shared_libs, and libdl depends on libd_static
via whole_static_libs (which means the latter's symbols are in
libdl.so). libc.so needs to link symbols defined in libdl.so, but before
this change, libdl.so was not threaded as a dependency of libc.so's
root's dependencies.

This CL adds dynamic_deps to cc_library_static and adds dynamic_library
files from those deps' CcSharedLibraryInfo into linker inputs. Now,
libdl.so and ld-android.so are correctly in libc.so's linker inputs.

This is needed to link libc.so successfully.

Test: TH
Fixes: b/187746106
Change-Id: I9f026f65b9715dcb0a106282b43d9e7c01bc8d69