bp2build: add stripped library support.
This CL adds an intermediate target in full_cc_library.bzl to optionally
strip a cc_shared_library when requested. It tries to mimic the logic in
cc/soong/strip.go as closely as possible.

There's also a simple smoke test that compares the file sizes of the
stripped libc.so with the original one.

Test: //build/bazel/tests:compare_libc_stripping
Test: TH
Change-Id: Ieef09817b991c3b113707d3bdc292326ed7dd4ba