Add a java binary, a java library, and an android app target building to bp2build ci.
Excluding cc jni targets as they currently don't compile (but the app does).
Adding the --fat_apk_cpu=k8 flag to common.bazelrc as it only applies to android app building rules, and it is currently required to successfully build them.
Forcing the use of the checked in Java JDK to avoid falling back to the local machine's JDK.

Test: ./build/bazel/ci/bp2build.sh runs successfully
Change-Id: I0f107d334f00c6ac7d91396a70c5023e48b6742b
Bug: 195786204
Bug: 195786195
Bug: 194639753