bp2build-progress: automatically generate list of converted modules.
This change calls `m bp2build` to generate the Soong injection workspace
and the list of converted modules as a static list, so there's no need
to manually maintain the list.

It also hardcodes the lunch product in the graph/bp2build generation so the
module graph is hermetic from the external running shell.

It also adds the list of converted modules to the end of the report.

Sample report: https://gist.github.com/jin/eaf722ca707602f1c3c797322fc72f79

Fixes: 199837056
Test: bp2build-progress.py report adbd
Change-Id: I1b4cd2c311509af9344d5ec6ced4b4fef797207d