Re-enable and fix bazel_apexer_wrapper_test.
This test was accidentally skipped by bp2build-incremental because its
dependency, //external/avb:avbtool, is a python binary that is target
incompatible with Android targets. Bazel automatically skips tests in a
target pattern like `...` if the test or any of its dependencies are not
compatible.

This CL makes it such that this test only runs when --platforms is the
linux host, and also add it to bp2build.sh for CI coverage.

Test: CI/bp2build-incremental
Test: b test //build/bazel/rules/apex:bazel_apexer_wrapper_test --platforms=//build/bazel/platforms:linux_x86_64
Change-Id: I7f0bec8b10c114fc4fb8b609921eaa636ca74763