Propagate tags for cc and sh rules, and make internal targets manual.
This allows internal targets to be always excluded from target patterns
(...), which is useful when using them for analysis only tests, because
some of these rules are not intended to be be buildable in the execution
phase.

The top level target will use the intended tags list from the callsite.

Test: CI
Change-Id: Iaabe58cd7555874c542974d31e7537505b6d5f3a