Remove min_sdk_version build setting
It was only read in the aspect, which can already
get it directly from the rule that invokes the
aspect.

It would probably also be more ideomatic to replace
the "apex_name" build setting with a "extra_cflags"
one that's owned by the cc rules, but I figured
I'd just stick with this for now.

Test: b build //build/bazel/examples/apex/minimal:build.bazel.examples.apex.minimal
Change-Id: Ibe484fd3253bb0e7a8085f55d2a182f9ff46be5e