Avoid symlinking shared lib output
This also propagtes a non-symlinked file in a new provider for
consumption in cquery. This, grouped together with a cquery handling
change in Soong, fixes a bug where shared libraries were showing up as
symlinks in a mixed build system image.

Test: With related soong changes, manually verified that installed
mixed-build libc was not a symlink. (cf_x86_64, inspected system.img
with `m droid`)
Test: mixed_libc.sh

Change-Id: I3dfb941d46e7aa8e45fc9985fe0d49886d830711