Support proto.include_dirs
For each package in the include_dirs property a proto_library target
should be added to the BUILD file in that package and a mapping
should be added to the bp2build code, by this way a proper dependency
relationship can be established and used by bazel.

Bug: 239944064
Test: Manually verified include_dirs can be
properly converted to bazel and used by bazel to build the targets.
Change-Id: I83f1877460e863e32ff1fb000eee0347d985fc71

Change-Id: Ibf3e9f8b1d57dfbc2e444fe7713c0928a6020fdc