Add Starlark support for cc_test targets.
    $ b test //packages/modules/adb:adb_test //system/core/... --config=linux_x86_64
    //packages/modules/adb:adb_test                                 (cached) PASSED in 13.2s
    //system/core/libcrypto_utils/tests:libcrypto_utils_test        (cached) PASSED in 0.1s
    //system/core/libcutils:libcutils_test                          (cached) PASSED in 0.1s
    //system/core/libutils:libutils_test                            (cached) PASSED in 5.2s
    //system/core/property_service/libpropertyinfoserializer:propertyinfoserializer_tests (cached) PASSED in 0.1s

This CL is an initial implementation of the cc_test macro, which reuses
a lot of cc_binary internals but generates a cc_test top level target
instead.

With this CL, about 50 cc_test targets can be built for multiple host
and device architectures (along with their data deps), as well as `bazel
test` can be used for those that are host_supported (implemented
through target_compatible_with bp2build conversion).

The cc_test macro is responsible for adding implicit copts and deps
based on the gtest attribute. I figured that a single gtest attribute is
safer for users instead of a disparate set of deps and copts, which can
get messy quickly.

I had to modify bp2build.sh to split out build and test commands because
now there are cc_test targets built for the device, and hence cannot be
invoked with `bazel test` on the host directly.

Test: presubmits
Test: b test //system/... --config=linux_x86_64 --verbose_failures
--test_output=errors --keep_going --build_tests_only

Fixes: 242038415

Change-Id: I24cd326eacb22daa9002be58efb703d909c77c9a