Add config settings for the power set of arch features
In soong, if multiple features match the current configuration, they're
all used. In bazel, we have to have unambiguous select() statements, so
we can't have two features that are both active in the same select().
One alternative is to split out each feature into a separate select(),
but then it's difficult to support exclude_srcs, which may need to
exclude things from the regular arch select() statement if a certain
feature is active. Instead, keep the features in the same select
statement as the arches, but emit the power set of all possible
combinations of features, so that bazel can match the most precise one.

Bug: 189972518
Test: ./build/bazel/ci/bp2build.sh
Change-Id: I2b30426634d1ce6d227c75bc3e83188456b77393