Prevent apex targets from trying to resolve toolchains for non-android target platforms, like linux_x86_64.
This CL adds an incoming transition to apex and partition rules to
ensure that they're always built for android. It does not make sense to
build them for other platforms because their toolchains don't support
non-android targets.

With the transition, regardless of what the parent configuration of apex
targets are, the incoming transition to the android platform will ensure
that the toolchain can be resolved. The transition sets the default
platform to the same one set in the bazelrc.

This CL also makes it possible to run `b test //build/bazel/...` for all
android and linux_x86_64 platforms, and relying on incompatible target
skipping to only execute supported tests.

Test: b test //build/bazel/...
Fixes: b/217756861

Change-Id: I8befa1e85a555e5c64ac6e4896b4ddd137890e8b