Avoid clobbering outputs when switching between products for better incremental builds.
This sets a unique --output_base for each $TARGET_PRODUCT in the
bp2build CI loop.

We could also make it such that bazel.sh sets the output user root based
on $TARGET_PRODUCT.. but I suspect that people don't switch products as
often as this script does in the loop. So let's optimize the CI case first.

Test: run the script
Test: presubmit
Change-Id: I4ffbebb35a98fd2806ea83474a5f3716bfa2cf45