Make outputs from gensrcs package-agnostic
Previously, gensrcs delares the outputs adjacent to the corresponding inputs. With this limitation, the inputs can't be from external packages and `framework-cppstream-protos` can't be converted to Bazel.

This CL makes adjustment to the paths of the outputs so that the srcs can be from any external package. This is more aligned with how it works in Soong.

Bug: 247784159
Test: CI
Change-Id: I855d8ab7e153e0d8e474db3682bc0bec2df39119