Share default android_app_certificate logic between android_binary and apex.
This CL adds support for apexes that don't define an explicit
certificate module dep. In these cases, the macro relies on default app
cert dir product variable to look up either the test key (if not
specified), or a checked in key with if the is string name
specified (with certificate_name).

Test: added various unit tests.

Bug: 249089160
Fixes: 249089160
Change-Id: Ie206d73dd05fbaae5a57a6fe44641da547ce239e