apex_key: support default cert dir lookup.
The Soong implementation of apex_key supports looking up the named key
in the DefaultAppCertificate directory if the product variable is specified.

To disambiguate named files from named modules, the bp2build converter
checks if the string is a module, and if so, sets the
{private,public}_key attr. Otherwise, set the {private,public}_key_name attr.

This is similar to what was done for android_app_certificate in aosp/2238695.

The Soong code also validates that the basename minus extension are the
same between the keys, so add that to the rule impl as well.

See Soong impl: https://cs.android.com/android/platform/superproject/+/master:build/soong/apex/key.go;l=75-86;drc=d079e0b2708ce0f4cce470db929f28dd8d2b0e80

Test: presubmits, new tests
Change-Id: Ic6e0473a87c9ff0ef01dd2d7047eaf728d423786