ensure clang-tidy runs on correct source files
Some object files (bcm_object.o) are passed to targets in the srcs
attribute, and clang-tidy was unsuccessfully trying to run on the action
that generates it (it was treating it as a typical compile action, when
in fact the action was a linking of assembly files).

Some C files are passed to Bazel targets in the srcs attribute instead
of the srcs_c attribute, and clang-tidy attempted to validate these
using the C++ flags rather than the C flags.

Bug: 195029134
Test: hardcode enabling tidy actions && b build //packages/modules/adb:adbd
Change-Id: I04257b5fd4c8e7f63ef4aac42f70afd56e309ac5