Handle aml/ndk configs in Bazel
AML and NDK configs are special product configs that specify
TARGET_ARCH_SUITE rather than a TARGET_ARCH.
Logic in make guarantees that the two cannot be set simultaneously:
https://cs.android.com/android/platform/superproject/+/master:build/make/core/board_config.mk;l=266;drc=06c6273073526f2f2b9cb4a813ea80b3d8e6a145

Previously, Bazel was only looking at DeviceArch (corresponding to
TARGET_ARCH) to determine the architecture Soong would consider the
"primary" for android. However, for targets: ndk and mainline_sdk, this
would fail because they specify TARGET_ARCH_SUITE instead. In these
cases, the architectures are hard-coded in Soong:
https://cs.android.com/android/platform/superproject/+/master:build/soong/android/arch.go;l=1690-1709;drc=c55b58692dc39ca699cdebb6a444b46f555e8a52
and explicitly set as the android targets in the config:
https://cs.android.com/android/platform/superproject/+/master:build/soong/android/config.go;l=445-468;drc=b45a2ea782074944f79fc388df20b06e01f265f7

Here, we recreate the logic in Soong but only to determine the _primary_
architecture for android, b/258802089 is open to determine how to handle
the other architectures, which are currently ignored.

* if ndk is specified, use ndk arch config
* otherwise, if aml is specified, used aml arch config
* finally, use the DeviceArch setting

We add an error if none of these are set to aid in debugging if
additional alternative specifications of target arch are supported.

Test: m ndk-eng; b build //bionic/libc --config=android
Change-Id: I7fb47ac6f11cc0bb23a19e19e4dbf677aaa86cd7