Delete default_android_transition
This transition was added so that the tests are run under
an android configuration.

In theory, instead of this, we should be able to mark apexes
as target_compatible_with = ["//build/bazel/platforms/os:android"],
and then the tests will also be treated as incompatiable as well, given
this docs quote:

"Targets that transitively depend on incompatible targets are themselves
considered incompatible. They are also skipped for building and testing."
https://bazel.build/reference/be/common-definitions#common.target_compatible_with

However, it appears that toolchains are still attempted to be resolved
even when the target being built is incompatible. To work around this,
make the apex toolchain optional.

Bug: 249685973
Test: b test --toolchain_resolution_debug='//build/bazel/rules/apex:apex_toolchain_type' //build/bazel/tests/apex:all
Change-Id: Ifa4b72f6d391094f5a779b44ccc720123589ffd8