Implement apex_available_aspect.
This aspect visits the payload-contributing deps of an apex, and
validates that certain prerequisite deps opt-in to be included in the
top-level apex by specifying apex_available=<name of apex> or
apex_avaiable=//apex_available:anyapex in its tags.

The tags are converted directly from the apex_available prop of the
original Soong module.

Test: presubmits
Fixes: 218841706
Change-Id: I0901e2a81aed7e6aeb04fbeb3ec2e89ad503bceb