Support next_version in aidl_interface macro
aidl_interface in Soong has the concept of next version which depends on the unstable prop and the versions specified in versions and versions_with_info props.

This commit adds support for next version so that upstream modules can import the +1 version of the aidl interface. When unstable is set in an aidl_interface, its next_version is "".

Bug: 261139208
Test: b build //frameworks/av/apex:com.android.media.swcodec --config=android
Change-Id: I6a08f632e4c16a7cd95e70c1d3dbc6126422683b