Add action to verify .hash file in aidl_interface frozen version
Semantically, an aidl_library represents either a group of aidl files or a frozen version of an aidl_interface. If an aidl_library represents a n aidl_interface's frozen version, version and hash_file attributes are set and hence result in the validation action AidlHashValidation.

This is to replicate a validation we have in Soong.
https://cs.android.com/android/platform/system/tools/aidl/+/master:build/aidl_api.go;l=46;drc=87bcb923b4ed9cf6e6837f4cc02d954f211c0b12

Test: b build //frameworks/hardware/interfaces/stats/aidl:android.frameworks.stats-V1
Test: Regression test: change the .hash file and run the b build command again. Expect to see a failure with the error message from message_check_integrity.txt
Change-Id: Ic755e01ea406c98e8e3f64509aaa7ffeef45d181