Add filesInfo data into ApexMkInfo provider.
This change adds a new ApexMkInfo.files_info field, containing a list of
string->string dicts. Each dict contains metadata about every entry in
the APEX's payload. It's used to reconstruct filesInfo in the mixed
builds Soong APEX handler, which is then used to generate Make
definitions/modules for compatibility with Make packaging rules.

We will need to do this for as long as there are packaging rules that
operate on the contents of an apex, like for the coverage symbol files.
This will need to be extended to other types of payload contents as
well, like apps and jars.

Bug: 271423316
Bug: 271423062
Test: presubmit
Test: new analysis test.
Change-Id: Ia4f0167bbb4022e0416baf5017074544ad49a633