Switch away from local_jdk (to jdk17)
the @local_jdk repository does not support platform-restrictions,
which will become an issue when needing to switch java toolchains
based on platform. In addition, there's a bit of inconsistency
currently between usages of jdk17 and jdk11, as the default java home
is set to jdk17 but we use some custom targets in prebuilts/jdk/jdk11.

This CL has bazel stop using the local_jdk and instead specify a
local_java_runtime under prebuilts/jdk/jdk17, which we'll be able to
impose platform restrictions on later.

The Kotlin rules directly refer to @local_jdk//jar, so a stub
local_jdk repository is defined to keep this reference valid.

Test: presubmits
Bug: 215230098
Change-Id: Ifcf443ab9cf69bcc95a10ace2e2bba951b085cf3