generate toc from final shared library file
In Soong, the TOC is generated from the final shared library file, and
some rules depend on the library file always being present if the TOC is
present. This change ensures for the Bazel rules that if a TOC is
present that the final shared library will also be present.

The shared_library_toc rule was removed since the TOC now depends on an
internal action action of the _cc_library_shared_proxy rule.

Test: m crosvm
Bug: 275416061
Change-Id: I8417f4831c4671f3ac90fa263c1add2a823ba761