Move the testing platforms out of soong_injection
I realized that we don't need to create all platforms in soong_injection
if we move the product_vars target out of soong_injection as well.

product_vars needs to be a select statement from all the products
to their _product_vars target, so we require all android_product()s
to "register" with that product_vars select statement.

Bug: 269577299
Test: b test --config=android //build/bazel/rules/apex:all
Change-Id: I47456ee1dbcc62af2c11cbff7e2b08152506eb29