Read the default certificate in a transition-friendly way
The default certificate will now be read from
//build/bazel/rules/product_config:product_vars.

However, now we can't add a dependency on the appropriate certificate
file at load time. All rules that use default certificates will have
to have dependencies on all certificates in the default certificate
folder, and then pick the correct one at analysis time. This problem
is further compounded by the fact that we can't glob() files outside
of our package, so we'll have to check in a
"android_certificate_directory" filegroup in each default certificate
directory. The product_vars target will depend on the appropriate
filegroup and re-export its files as DefaultAppCertificateFiles.

Bug: 269577299
Test: b build --config=android //build/bazel/examples/apex/certificate_name:build.bazel.examples.apex.certificate_name on both an aosp product which doesn't set PRODUCT_DEFAULT_DEV_CERTIFICATE, and an internal product which does.
Change-Id: I0baf62cc32e8be3cd9e216250e94bcc8e6d5c72d