Test for differing bp2build files in multiproduct_analysis
bp2build should always produce the same build files for the same source
tree, regardless of which product we have lunched. Add a test for that.

Currently, there are bugs in bp2build that cause it to not produce
the same build files all the time, so we can't make this a failure yet.

Bug: 247785938
Test: ./multiproduct_analysis.sh
Change-Id: I352bedf3aa4ce6fcb3e50e63b40b944cd58b9937