Fix non-determinism of .tar.gz files
On CI, the .tar.gz files of the bp2build files have a timestamp in them.
Locally, they don't for some reason. I think glinux might be configured
somehow to have gzip not output timestamps, but the build servers
aren't.

Tar does not provide an option to not add a timestamp to the gzip
header, so instead gzip the file seperately.

Bug: 247785938
Test: Run this cl on go/abtd first and verify that there are fewer products that have differing bp2build files
Change-Id: I7635a6d7cf81c23a11e7836676eaf81001f52730