Fix non-determinism of .tar.gz files
On CI, the .tar.gz files of the bp2build files have a timestamp in them.
Locally, they don't for some reason. I think glinux might be configured
somehow to have gzip not output timestamps, but the build servers
aren't.

Tar does not provide an option to not add a timestamp to the gzip
header, so instead gzip the file seperately.

Bug: 247785938
Test: Presubmits
Change-Id: Ic7c709d9ab6b89f3a02a5a59d1c2349c16caf765