Don't load product_variables.bzl where possible
product_variables.bzl should not be loaded directly, but instead
read from the transitionable //build/bazel/product_config:product_vars
target.

Some of the information in product_variables.bzl is constant, and
doesn't need to be read from that target. In those cases, load the
information from other files so that the product_variables.bzl file
can be made private.

Bug: 269577299
Test: m nothing
Change-Id: I69d9d5479b51ff9098a3ff8f2caf5daa9fb4d49a