Add starlark data schema validator
Validating that input is correct in starlark is tedious and error-prone,
it's easy to forget to validate things like types, valid values,
extra keys, etc. You essentially have to write a denylist of all the
things that aren't allowed. It would be easier if you could write an
allowlist of valid data, which is what the schema validator is for.
You can write a schema, and then the validator will check that the data
exactly matches the schema without extra stuff.

The motivating usecase is for build/make/core/release_config.bzl,
which is essentially all validation of the release config flags.
It was refactored and introduced bugs in the validation, so handcrafting
the validation isn't working well there.

Test: b test //build/bazel/utils:type_validation_test_suite
Change-Id: Iefdd00702abbd47cc55c638ae607e84670948d39